+++ contracts/KyberNetworkProxyInterface.sol
@@ -3,20 +3,18 @@ pragma solidity ^0.5.7;
 import "./ERC20Interface.sol";
 
 interface KyberNetworkProxy {
-
     function trade(
         ERC20 src,
-        uint srcAmount,
+        uint256 srcAmount,
         ERC20 dest,
         address destAddress,
-        uint maxDestAmount,
-        uint minConversionRate,
+        uint256 maxDestAmount,
+        uint256 minConversionRate,
         address walletId
-        ) external payable returns (uint);
+    ) external payable returns (uint256);
 
-    function getExpectedRate(
-        ERC20 src,
-        ERC20 dest,
-        uint srcQty
-        ) external view returns (uint, uint);
+    function getExpectedRate(ERC20 src, ERC20 dest, uint256 srcQty)
+        external
+        view
+        returns (uint256, uint256);
 }
